requirements.txt
together
streamlit
python-dotenv
watchdog
